Add boilerplate to match registries project convention
