+++ docker-compose.yml
@@ -12,15 +12,13 @@ services:
 
   db:
     image: postgres:11
-    ports:
-      - 5432:5432
     networks:
       - bcros
     volumes:
       - postgres-data:/var/lib/postgresql/data
     networks:
       - bcros
-  
+
   search_api:
     image: search_api
     build:
@@ -32,7 +30,7 @@ services:
     environment:
       - FLASK_ENV=development
       - FLASK_DEBUG=1
-      - PYTHONPATH=/opt/app
+      - PYTHONPATH=/opt/app/src
     networks:
       - bcros
