+++ search-api/.s2i/environment
@@ -1,0 +1,1 @@
+APP_CONFIG=gunicorn_config.py
