+++ search-api/gunicorn_config.py
@@ -11,12 +11,14 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Provides the WSGI entry point for running the application
+"""The configuration for gunicorn, which picks up the
+   runtime options from environment variables
 """
-from search_api import create_app
 
-# Openshift s2i expects a lower case name of application
-application = create_app() # pylint: disable=invalid-name
+import os
 
-if __name__ == "__main__":
-    application.run()
+workers = int(os.environ.get('GUNICORN_PROCESSES', '1'))  # pylint: disable=invalid-name
+threads = int(os.environ.get('GUNICORN_THREADS', '1'))  # pylint: disable=invalid-name
+
+forwarded_allow_ips = '*'  # pylint: disable=invalid-name
+secure_scheme_headers = {'X-Forwarded-Proto': 'https'}  # pylint: disable=invalid-name
