+++ search-api/requirements/dev.txt
@@ -1,0 +1,22 @@
+
+# Everything the developer needs in addition to the production requirements
+-r prod.txt
+
+coverage
+pylint
+pylint-flask
+pytest
+pytest-cov
+pytest-env
+pytest-dotenv
+pytest-mock
+flake8
+flake8-blind-except
+flake8-debugger
+flake8-docstrings
+flake8-isort
+flake8-quotes
+pep8-naming
+autopep8
+pydocstyle
+FreezeGun
