+++ search-api/requirements/repo-libraries.txt
@@ -1,15 +1,1 @@
-psycopg2
-Flask-SQLAlchemy==2.4.1
-Flask==1.1.1
-SQLAlchemy-Continuum==1.3.9
-SQLAlchemy-Utils==0.34.2
-SQLAlchemy==1.3.10
-flask-marshmallow==0.10.1
-marshmallow-sqlalchemy==0.19.0
-marshmallow==2.20.5
-psycopg2-binary==2.8.4
-xlrd==1.2.0
-xlwt==1.3.0
-Flask-Migrate
-flask-cors
-openpyxl==3.0.3
+-e git+https://github.com/bcgov/sbc-common-components.git#egg=sbc-common-components-2.0.0&subdirectory=python
