+++ search-api/scripts/verify_license_headers.sh
@@ -1,3 +1,5 @@
+#!/usr/bin/env bash
+
 # Copyright © 2019 Province of British Columbia
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -11,12 +13,18 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Provides the WSGI entry point for running the application
-"""
-from search_api import create_app
 
-# Openshift s2i expects a lower case name of application
-application = create_app() # pylint: disable=invalid-name
 
-if __name__ == "__main__":
-    application.run()
+COPYRIGHT="Copyright © 2019 Province of British Columbia"
+RET=0
+
+for file in $(find $@ -not \( -path */venv -prune \) -not \( -path */migrations -prune \) -not \( -path */tests -prune \) -not \( -path */.egg* -prune \) -name \*.py)
+do
+  grep "${COPYRIGHT}" ${file} >/dev/null
+  if [[ $? != 0 ]]
+  then
+    echo "${file} missing copyright header"
+    RET=1
+  fi
+done
+exit ${RET}
